Add files via upload
The function definition takes 3 things: the array, the size, and the starting index i.

In main, I called it with 2 things: findMaxRecTail(myArray, SIZE).

That only works if the compiler already knows that the 3rd thing (i) has a default value (= 0) at the moment of the call.

If the compiler hasn’t seen the prototype with int = 0 before the call, it assumes there’s no default, so it complains: “expects 3, got 2.”

That's why I was having problems with code! To fix that I pass the third as an arguement, I just added 0 to the two fuinctions on main.